requirements/common.txt
psycopg2==2.6.2
appdirs==1.4.0
backports.shutil-get-terminal-size==1.0.0
certifi==2017.1.23
decorator==4.0.11
Django==1.10.5
django-debug-toolbar==1.5
django-modeladmin-reorder==0.2
django-polymorphic==1.1
djangorestframework==3.5.3
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
PyYAML==3.12
simplegeneric==0.8.1
simplejson==3.10.0
six==1.10.0
sqlparse==0.2.2
traitlets==4.3.1
urllib3==1.20
wcwidth==0.1.7
